move pages to separate folder
