+++ Ghost.Web.React/src/pages/Genres.jsx
@@ -2,15 +2,15 @@ import { Chip } from '@mui/material'
 import React, { useState, useEffect } from 'react'
 import { fetchGenres } from '../services/genre.service'
 import { Link } from 'react-router-dom'
-import { ChipSkeleton } from './ChipSkeleton.jsx'
-import { NothingHere } from './NothingHere.jsx'
-import { Search } from './Search'
+import { ChipSkeleton } from '../components/ChipSkeleton.jsx'
+import { NothingHere } from '../components/NothingHere.jsx'
+import { Search } from '../components/Search'
 import { Box } from '@mui/system'
 import { includes } from 'ramda'
 import usePromise from '../services/use-promise'
 
 export const Genres = () => {
-  const [genres,, loadingGenres] = usePromise(() => fetchGenres())
+  const [genres, , loadingGenres] = usePromise(() => fetchGenres())
   const [filteredGenres, setFilteredGenres] = useState([])
   const [search, setSearch] = useState('')
