+++ Ghost.Web.React/src/pages/Home.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { fetchVideos, fetchRandomMedia } from '../services/video.service'
-import { VideoView } from './VideoView.jsx'
+import { VideoView } from '../components/VideoView.jsx'
 
 export const Home = () => {
   return <VideoView fetchFn={fetchVideos} fetchRandomVideoFn={fetchRandomMedia} />
