+++ Ghost.Web.React/src/pages/Jobs.jsx
