+++ Ghost.Web.React/src/pages/Media.jsx
@@ -9,22 +9,22 @@ import FirstPageIcon from '@mui/icons-material/FirstPage';
 import LastPageIcon from '@mui/icons-material/LastPage';
 import TheatersIcon from '@mui/icons-material/Theaters';
 
-import { Video } from './Video.jsx'
-import { VideoGenres } from './VideoGenres.jsx'
-import { VideoActors } from './VideoActors.jsx'
-import { TextEdit } from './TextEdit.jsx'
-import { VideoMetaData } from './VideoMetaData.jsx'
-import { items, VideoMenu } from './VideoMenu.jsx'
-import { ChipSkeleton } from './ChipSkeleton.jsx'
-import { Chapters } from './Chapters.jsx'
-import { generateVideoUrl, toggleFavourite } from '../services/video.service'
-import { FavouriteIconButton } from './FavouriteIconButton.jsx'
+import { Video } from '../components/Video.jsx'
+import { VideoGenres } from '../components/VideoGenres.jsx'
+import { VideoActors } from '../components/VideoActors.jsx'
+import { TextEdit } from '../components/TextEdit.jsx'
+import { VideoMetaData } from '../components/VideoMetaData.jsx'
+import { items, VideoMenu } from '../components/VideoMenu.jsx'
+import { ChipSkeleton } from '../components/ChipSkeleton.jsx'
+import { Chapters } from '../components/Chapters.jsx'
+import { generateVideoUrl, toggleFavourite } from '../services/video.service.js'
+import { FavouriteIconButton } from '../components/FavouriteIconButton.jsx'
 import usePromise from '../services/use-promise.js'
-import { ProgressIconButton } from './ProgressIconButton.jsx'
-import { MediaSection } from './MediaSection.jsx'
-import { VideoCard } from './VideoCard.jsx'
-import { AddVideoCard } from './AddVideoCard.jsx'
-import { TextInputModal } from './TextInputModal.jsx'
+import { ProgressIconButton } from '../components/ProgressIconButton.jsx'
+import { MediaSection } from '../components/MediaSection.jsx'
+import { VideoCard } from '../components/VideoCard.jsx'
+import { AddVideoCard } from '../components/AddVideoCard.jsx'
+import { TextInputModal } from '../components/TextInputModal.jsx'
 import { fetchMedia } from '../services/media.service.js'
 
 const fetchGenres = async (id) => (await axios.get(`/genre/video/${id}`)).data
